use 64bit unsigned integer for iterations in paint use -pg -no-pie on all binaries

+++ tp1-artist-ant/src/ant_engine.h
@@ -12,7 +12,7 @@
     *(p) = (v) % (limit);   \
   } while(0)             
  
-void *paint(void *artist_ant, void *gridfn, colour_fn next_colour, rule_fn rotation_for_colour, uint32_t iterations);
+void *paint(void *artist_ant, void *gridfn, colour_fn next_colour, rule_fn rotation_for_colour, uint64_t iterations);
 
 orientation_t new_orientation(orientation_t orientation, rotation_t rule);
